% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clrscatter.R
\name{clrscatterplot}
\alias{clrscatterplot}
\title{Plots a scattergram of two clr transformed variables.}
\usage{
clrscatterplot(comp, xval, yval, labels = FALSE)
}
\arguments{
\item{comp}{Object of type compositional}

\item{xval, yval}{String referring to the names of the variables on x and y axes}
}
\description{
The nature of compositional data means that plotting raw percentages against eachother on a scatterplot
is fraught with problems. This can be improved by transforming the variables into clr space. Ignore the warnings.
}
\details{
Pearson, K., 1897. Mathematical contributions to the theory of evolution.—On a form of spurious correlation which may arise when indices are used in the measurement of organs. Proc. R. Soc. Lond. 60, 489–498. https://doi.org/10.1098/rspl.1896.0076
}
\examples{
\dontrun{clrscatterplot(qem_rep,"Diagenetic","Porosity")}
}
\seealso{
clr, makePCAbiplot, load_compositional
}
\keyword{clr,}
\keyword{compositional,}
\keyword{scatterplot}
