% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributional.R
\name{load_distributional}
\alias{load_distributional}
\title{Loads in distributional data for analysis}
\usage{
load_distributional(datafilename, tag, sample_id, val_id, type_id,
  typing = NULL, err_id = NULL, SH = FALSE)
}
\arguments{
\item{datafilename}{Path and filename of the .csv file containing the data}

\item{tag}{A string to identify the distributional object for multiproxy analyses e.g. INDSCAL}

\item{sample_id}{String which corresponds to the column in the input file which contains the sample names.
Note that this is the name of the sample (of which there are many individual measurements), NOT the id
of each individual measurement.}

\item{val_id}{A string corresponding to the column which has the variable (e.g. Age, temperautre etc..)}

\item{type_id}{String corresponding to column which contains the different classification (e.g. Formation, Age Period, protolith)}

\item{typing}{Path and filename of the .csv file containing the colour scheme/typing. This should have a column "Colour" and "Category". The Colour
column should contain hexadec codes of colours, and the Category should contain the identifier of the group
that each colour should correspond to. These identifiers should be contained in the input file under the column header
given by "type_id".}

\item{err_id}{optional String corresponding to column which contains the 2S errors for variable. Required for Sircombe-Hazleton dissimilarity}

\item{SH}{optional Boolean indicating whether the default dissimilarity measure should be the Sircombe-Hazleton distance.
The default is FALSE indicating that the Kolmogorov-Smirnov distance is used.}
}
\value{
Returns an object of type "distributional".
}
\description{
This loads data into the "distributional" data class of the "provenace" package. Distributional data
is that which for each sample it is the "distribution" of data which holds information. This
includes any "spectrum", e.g. UPb ages etc... The "distributional" object this generates can
be used in any of the default functions in "provenance" as well as those defined in this
package.
The input file which this reads is a .csv in which each measurement is a row, and the columns
correspond to the variable, classification, and sample name. Note that this is a different format
to the format required for the "provenance" package.
}
\details{
Typing: Provenance analysis is broadly about categorising data into different "groups". This is best illustrated
by example, see the worked examples in the vignettes.
This approach allows different groups to be intuitively compared by fixing predefined groups to a particular
colour when the samples are loaded. This colour scheme is the same for every subsequent analysis allowing
for easier comparison. For example, it allows samples from one particular formation to be coloured the
same in UPb age spectra, MDS plots or PCA plots. Different groups of samples (e.g. Formation) are specified in a seperate file
which is simply a spreadsheet of the specific types (e.g. Old Red Sandstone, Greenstone) with a corresponding hexadecimal colour code.
This has to be manually inputted only once for the entire dataset. If this is not done, the automatically
created colour scheme will not be consistent, and could repeat colours, which makes it more difficult to compare
different groupings from method to method. A useful tool to generate hexadecimal codes from colours is: http://htmlcolorcodes.com/.
}
\examples{
\dontrun{zircs <- load_distributional("zircondata.csv","typing_palette.csv","Zircon Ages","SampleNo", "PreferredAge", "SandType")}
\dontrun{zircs <- load_distributional("zircondata.csv","typing_palette.csv","Zircon Ages","SampleNo", "PreferredAge", "SandType", "PreferredAge_2s",SH = TRUE)}
}
\seealso{
read.distributional
}
\keyword{distributional}
