% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makemultiproxplots.R
\name{makeindscalmap}
\alias{makeindscalmap}
\title{Makes an INDSCAL map for multiple analyses on the same data-set.}
\usage{
makeindscalmap(..., typing = NULL, labels = FALSE)
}
\arguments{
\item{...}{Multiple objects either of class compositional or distributional or both. For this function to work
the data-sets must contain overlapping samples. For example a data-set which has mineralogy, major elements and UPb ages for each sample.}

\item{typing}{An object of class distributional (see "load_distributional"), e.g. Zircon ages, rutile ages etc...}
}
\value{
Returns a ggplot object. Automatically plots it, but can be saved to variable as any normal ggplot object.
}
\description{
See indscal in "provenanace" package documentation, this is effectively just a wrapper.
The dissimilarity measures used are inherited from the individual objects.
}
\details{
Vermeesch, P., Resentini, A., Garzanti, E., 2016. An R package for statistical provenance analysis. Sediment. Geol., Sediment generation and provenance: processes and pathways 336, 14-25. https://doi.org/10.1016/j.sedgeo.2016.01.009
}
\examples{
\dontrun{inds <- makeindscalmap(zircs,rus,rutmps,typing="typing_palette.csv")}
}
\keyword{INDSCAL}
\keyword{compositional,}
\keyword{distributional,}
