% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMDSmap.R
\name{makeMDSmap}
\alias{makeMDSmap}
\title{Make a Multi-Dimensional Scaling (MDS) map of data}
\usage{
makeMDSmap(obj, classical = FALSE, labels = FALSE)
}
\arguments{
\item{obj}{An object either of class compositional or distributional (see "load_compositional" and "load_distributional")}

\item{classical}{Boolean, defaults to False. Indicates whether MDS should be non-metric or classical.}

\item{labels}{Indicates whether labels should be plotted for samples or not. Can be slow for large data-sets...}
}
\value{
Returns a ggplot object. Automatically plots it, but can be saved to variable as any normal ggplot object.
}
\description{
This is effectively a wrapper for the MDS function in "provenance". Can perform both
non-metric and classical MDS. Returns the "stress" of the MDS fit, to assess the goodness of fit.
}
\examples{
\dontrun{makeMDSmap(zircs,labels=T)}
}
\keyword{MDS,}
\keyword{compositional,}
\keyword{distributional}
