% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePCAbiplot.R
\name{makePCAbiplot}
\alias{makePCAbiplot}
\title{Make a principal component biplot of a compositional dataset}
\usage{
makePCAbiplot(compo, haxis = "Comp.1", vaxis = "Comp.2", rayscaler = 1,
  labels = FALSE)
}
\arguments{
\item{compo}{An object of class compositional (see "load_compositional"), e.g. mineralogy, major oxides etc... This cannot
contain zeros.}

\item{haxis, vaxis}{These are strings of form "Comp. n" where n is a number. These
are by default "Comp.1" and "Comp.2" and they indicate which components to plot on the
horizontal and vertical axes respectively. Note that at present there is a bug where the
variation displayed on the vertical axis corresponds to the axis which contains less variation
irrelevant of whether it is actually the v axis. To be safe, always plot the higher order component
on the vertical axis.}

\item{rayscaler}{Number which adjusts the length of the rays. default is 1, but vary to best
fit the data.}
}
\value{
Returns a ggplot object. Automatically plots it, but can be saved to variable as any normal ggplot object.
}
\description{
This function plots both samples and variables (hence the "bi" in biplot) graphically via Principal Component Analysis (PCA).
This is effectively a wrapper for the PCA function in "provenance", but visualises it
better, especially for large data-sets. Automatically displays components 1 and 2, but
can display further ones. Prints out the proportional variation on each component
so that it can be seen how many components are significant.
}
\details{
Aitchison, J., Greenacre, M., 2002. Biplots of Compositional Data. J. R. Stat. Soc. Ser. C Appl. Stat. 51, 375-392.
}
\examples{
\dontrun{makePCAbiplot(qem_rep,haxis="Comp.1",vaxis="Comp.2")}
}
\seealso{
maketernaryplot,load_compositional,replacezeros
}
\keyword{PCA}
\keyword{biplot,}
\keyword{compositional,}
