% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQQplot.R
\name{makeQQplot}
\alias{makeQQplot}
\title{Make a Lower Quartile - Upper Quartile plot of distributional data}
\usage{
makeQQplot(dist, labels = FALSE)
}
\arguments{
\item{dist}{An object of class distributional (see "load_distributional"), e.g. Zircon ages, rutile ages etc...}
}
\value{
Returns a ggplot object. Automatically plots it, but can be saved to variable as any normal ggplot object.
}
\description{
This function plots the lower quartile vs the upper quartile for a range of distributions after
Anderson et al (2017). Note this is not a typical
}
\details{
Andersen, T., Kristoffersen, M., Elburg, M.A., 2017. Visualizing, interpreting and comparing detrital zircon age and Hf isotope data in basin analysis - a graphical approach. Basin Res. https://doi.org/10.1111/bre.12245
}
\examples{
\dontrun{QQplot <- makeQQplot(zircs)}
}
\keyword{distributional,}
\keyword{quartile}
