% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeridgeplot.R
\name{makeridges}
\alias{makeridges}
\title{Make ridge plot of distributional data}
\usage{
makeridges(distributional, scale = 10, clustered = T, from = NA,
  to = NA, samebandwidth = F, normalise = T, adaptive = T)
}
\arguments{
\item{distributional}{An object of class distributional (see "load_distributional"), e.g. Zircon ages, rutile ages etc...}

\item{scale}{This is the degree of overlap for the ridges. A value of 1 means that the largest overlap in the
plot touches the base of the ridge above. Default value is 10, adjust to see what works best for individual cases.}

\item{clustered}{Boolean, defaults to TRUE indicates whether the samples are to be sorted vertically according
to the similarity of their spectra. If FALSE, they are grouped according to their "typing"}

\item{from, to}{Minimum and maximum x value. Optional}

\item{samebandwidth}{Boolean, default FALSE. If TRUE, all KDEs are generated with the same bandwidth (not recommended)}

\item{normalise}{Boolean, default TRUE. All KDEs given same area.}

\item{adaptive}{Boolean, default TRUE. Geological spectra are often multi-modal, and so the KDE algorithms
should have an "adaptive" bandwidth which varies across the x-axis according to point density. This is broad
when data is sparse but narrow when data is dense. Turn off only if data is strongly unimodal.}
}
\value{
Returns a ggplot object. Automatically plots it, but can be saved to variable as any normal ggplot object.
}
\description{
This function creates overlapping KDEs and plots them as an overlapping 'ridge plot'
inspired by the iconic "Unknown Pleasures" album cover by Joy Division. Whilst being very visually
attractive it also allows large amounts of KDEs to be plotted in one go, allowing for succint representation.
This also allows the different KDE's to be sorted by a clustering algorithm, which groups similar spectra
together according to their Kolmogorov-Smirnov statistic.
}
\details{
For large datasets this can take a while but be patient!
}
\examples{
\dontrun{ridgeplot <- makeridges(zircs)}
\dontrun{ridgeplot <- makeridges(zircs,samebandwidth = FALSE,  normalise = FALSE,  adaptive = FALSE)}
\dontrun{ridgeplot <- makeridges(zircs,clustered = FALSE, scale = 20)}
\dontrun{ridgeplot <-  makeridges(zircs, from = -500, to = 2500)}
\dontrun{ridgeplot <- makeridges(zircs, clustered = FALSE)}
}
\keyword{KDEs}
\keyword{distributional,}
\keyword{ridges,}
