% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternaryplot.R
\name{maketernaryplot}
\alias{maketernaryplot}
\title{Makes a ternary plot}
\usage{
maketernaryplot(comp, component1, component2, component3, centered = FALSE,
  size = 1, isoportions = TRUE)
}
\arguments{
\item{comp}{An object of type compositional. This should not contain any zeros.}

\item{component1, component2, component3}{Strings of the three components to plot against eachother}

\item{centered}{Boolean indicating whether the ternary phase diagram should be "centered" or not. This
is a technique which moves the geometric mean of a data-set to the centre of the plot, thus allowing
making variations at the edge of the space visible. see Von Eynatten et al. (2002).}

\item{size}{Value which adjusts sizes of the points}

\item{isoportions}{Boolean indicating whether isoPortion lines should be added. Recommended if the data
is centered as it allows for the true values to be read off.}
}
\value{
Returns a plot object.
}
\description{
Makes a ternary plot
}
\examples{
\dontrun{maketernaryplot(qem_rep,"Quartz","Plag","Kfsp",centered=TRUE,size=1,isoportions=TRUE)}
}
\seealso{
load_compositional, replacezeros
}
\keyword{compositional}
